+++ test/test.js
@@ -1,0 +1,8 @@
+/*!
+ * Copyright (c) 2020-2023 Digital Bazaar, Inc. All rights reserved.
+ */
+import * as bedrock from '@bedrock/core';
+import '@bedrock/credentials-context';
+import '@bedrock/test';
+
+bedrock.start();